+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - PR #1098 Add new graph classes to support 2D partitioning
 - PR #1124 Sub-communicator initialization for 2D partitioning support
 - PR #1147 Added support for NetworkX graphs as input type
+- PR #1157 Louvain API update to use graph_container_t
 
 ## Improvements
 - PR 1081 MNMG Renumbering - sort partitions by degree